tests/requirements-test.txt
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
fastapi.testclient==0.7.0